test: use scenario for unit testing (pt.2)
